Updated Visual Studio Solution and added new projects
This commit includes a significant update to the Visual Studio Solution, including an update to the format version and the addition of two new projects. The solution and project configurations have been expanded to include Debug and Release configurations for multiple platforms. A new SolutionGuid has also been added. Localization has been implemented for various languages in the `Add-AppDevPackage.psd1` file. Several PNG image files have been added or modified, but the specific changes to the images are not clear from the diff. The `Package.appxmanifest` file was updated with new application details and a new `Package.StoreAssociation.xml` file was added. The `project.assets.json` file has been updated to include a new version of the `Microsoft.Windows.SDK.BuildTools` package and a new project.