Ver.: 10.16.11(.89)
The most significant changes in the code include the update of the application version from 10.15.10 to 10.16.11, the correction of the application ID and namespace, the addition of a new project to the Visual Studio solution file, and the addition of several new Objective-C++ source files.

1. The application version has been updated from 10.15.10 to 10.16.11. This change is reflected in multiple files including `CMakeLists.txt`, `AndroidManifest.xml`, `GuitarFineTune.cbp`, `resources.rc`, `Makefile`, `project.pbxproj`, `Info-App.plist`, `GuitarFineTune.sln`, and `GuitarFineTune_App.vcxproj`. The software version has also been updated across multiple files including MajorMinorBuildRevisionNumbers.xml, resources.rc, project.pbxproj, Info-App.plist, GuitarFineTune.jucer, JuceHeader.h, Package.appxmanifest, guitarfinetune.desktop, and snapcraft.yaml.

2. The application ID and namespace have been corrected from "dk.eksit.guitarfinetunep" to "dk.eksit.guitarfinetune" in `build.gradle` and `AndroidManifest.xml`.

3. A new project "packProj" has been added to the Visual Studio solution file `GuitarFineTune.sln`. Also, the project section "MajorMinorBuildRevisionNumbersXmlFile" has been updated and the global section "ExtensibilityGlobals" has been removed from `GuitarFineTune.sln`.

4. Several new Objective-C++ source files (".mm" files) have been added to the GuitarFineTune_App.vcxproj.filters file, indicating that parts of the project are being enhanced or rewritten with Objective-C++ code, likely to improve macOS-specific functionality. These new files have been added to various filters within the project for better organization. Additionally, compilation instructions for new files related to iOS and MacOS from the JUCE library have been added, suggesting an expansion of cross-platform capabilities. Lastly, an extra closing tag for the project was removed from the file.

Other changes include the update of the Android version code from 87 to 89 in `CMakeLists.txt`, `AndroidManifest.xml`, and `GuitarFineTune.jucer`, the update of the SDK directory path in `local.properties`, the removal of the release signing configuration from `build.gradle`, and the update of the build number from "87" to "89" in MajorMinorBuildRevisionNumbers.xml, Info-App.plist, GuitarFineTune.jucer. The hexadecimal representation of the version number has also been updated from "0xa0f0a" to "0xa100b" in project.pbxproj and JuceHeader.h.